Added generate_outputs "test" plus moved fixtures so test could access them
